requirements.txt
numpy==1.17.5
matplotlib>=3.0.0
opencv-python>=3.4.0
setuptools
Cython
scipy>=1.2.0
six
h5py
ninja
yacs
tqdm
scikit-image
Pillow>=6.0.0
tabulate
pycocotools
torch>=1.6.0
torchvision>=0.7.0
